+++ src/main/java/edu/ksu/canvas/aviation/repository/AttendanceRepositoryCustom.java
@@ -9,7 +9,7 @@ import edu.ksu.canvas.aviation.entity.Attendance;
 public interface AttendanceRepositoryCustom {
 
 
-    List<Attendance> getAttendanceByCourseAndDayOfClass(int courseId, Date dateOfClass);
+    List<Attendance> getAttendanceByCourseAndDayOfClass(long courseId, Date dateOfClass);
 
     void saveInBatches(List<Attendance> attendances);
